+++ movie-smt-bff/app.go
@@ -1,38 +1,14 @@
 package main
 
 import (
-	themoviedbgateway "movie-smt-bff/themoviedb_gateway"
-	"net/http"
+	controllers "movie-smt-bff/controller"
 
 	"github.com/gin-gonic/gin"
 )
 
-type apiError struct {
-	Message string `json:"message"`
-}
-
-var nameNotSuppliedError = apiError{Message: "query parameter 'name' must be supplied!"}
-var internalServerErrorError = apiError{Message: "internal server error"}
-
 func main() {
-	r := gin.Default()
+	router := gin.Default()
+	controllers.Routes(router)
 
-	r.GET("/search", search)
-
-	r.Run()
-}
-
-func search(c *gin.Context) {
-	name := c.Query("name")
-	if name == "" {
-		c.JSON(http.StatusBadRequest, nameNotSuppliedError)
-		return
-	}
-
-	res, err := themoviedbgateway.SearchMovies(name)
-	if err == nil {
-		c.JSON(http.StatusOK, res)
-	} else {
-		c.JSON(http.StatusInternalServerError, internalServerErrorError)
-	}
+	router.Run()
 }
