+++ movie-smt-bff/service/movie_service.go
@@ -1,4 +1,4 @@
-package themoviedbgateway
+package service
 
 import (
 	"encoding/json"
